requirements.txt
Flask==3.0.3
flask-restx==1.3.0
flask-cors==4.0.1
huggingface_hub>=0.25.0
python-dotenv==1.0.1
gunicorn==21.2.0
